Add basic support for X11
This adds just enough code for Avizo to also work on X11.

The code to check whether Avizo is running on the Wayland backend is
copied from:
<https://mail.gnome.org/archives/vala-list/2016-December/msg00012.html>

GtkLayerShell code is wrapped around a check for a Wayland display,
otherwise the X11 code path is taken.

The X11 code path uses EWMH window type hints to prevent the window
from grabbing focus when spawning.